Move summary variable code to different file
